% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geo_clean.R
\name{Geo_clean}
\alias{Geo_clean}
\title{Geo_clean}
\usage{
Geo_clean(ISO, Level, dataset)
}
\arguments{
\item{ISO}{ISO3 code of your targeted country}

\item{Level}{Level of your targeted ADM shape file}

\item{dataset}{dataset by your own with NetCDF format}
}
\value{
A data frame/csv object that contains geodata for the specified time period.
}
\description{
Provides a CSV data with your own Geo data (That dowload from somewhere)
}
\examples{
data <- Geo_extract(ISO = "VNM", Level = "1", dataset = "Your nc path file")
}
