% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geo_plot.R
\name{Geo_plot}
\alias{Geo_plot}
\title{Geo_plot}
\usage{
Geo_plot(ISO, Level, dataset, palette)
}
\arguments{
\item{ISO}{ISO3 code of your targeted country}

\item{Level}{Level of your targeted ADM shape file}

\item{dataset}{dataset by your own with NetCDF format}

\item{palette}{palette color that you want for your map}
}
\value{
A map that displays geodata for the specified time period.
}
\description{
Provides a beautiful map with your own Geo data (downloaded from somewhere)
}
\examples{
map <- Geo_plot(ISO = "VNM", Level = "1", dataset = "Your nc path file", palette = "Blues")
}
